dht11 funciona ok

+++ wifi-mqtt/station/main/dht.h
@@ -5,7 +5,7 @@
 typedef struct 
 {
     float temperature;
-    float humidity; 
+    int humidity; 
     
 }dht;
